build: markdown and whitespace formatting
Add `react-markdown`, which wraps markdown formatted strings into a
`ReactMarkdown` component, so that `dangerouslySetInnerHTML` is not
used.

Globally sets the `white-space: pre-line` CSS so that `\n` new lines are
rendered in HTML. Whitespace formatting needs to go inside the <App />.
Otherwise, extraneous whitespace may render outside the app depending on
how the HTML is built.