style: add first page preview via <hr>
Added a styled <hr class="paper-size"/> to show the bottom margin and
end of first page. The element is border-box sized and positioned
absolute, relative to the <App /> parent container. Text is written
using the `::before` pseudoclass.

Media query is used to prevent the element from being visible when sent
to print.