fix(App): initialize App with Effect
Move the fetch inside the App component, and initialize using an Effect
that fetches the sample.

Fetching the sample with a top-level await to set it to the initial
state caused problems on build/export. Instead, using Effect inside the
component will hopefully be able to synchronize the resume state to the
public asset while still only running once, when the App component is
rendered.