feat(SideBar): section visibility control
Use a "SideBar" component to keep track of which sections are visible,
using `showSection` state variable. Eventually, the SideBar component
may truly reside on the side but, for now, it is just a second top bar.

Control the visibility of sections using JSX rather than CSS--
conditionally render the component after checking the value of the
`showSection` state variable .

Control the visibility of SideBar itself using a state variable, and add
a toggle to the NavBar component.