feat(Text): add a Text section for cover letters
Cover letters are expected to be markdown files. JSON resume fields
written into the front matter will be rendered before the text body.
Use case is including a Basics section to provide a letter head that
exactly matches a resume.

Add a data source controller to the sidebar: MarkdownSourceController
- Reads markdown files as text from URL or file input
- Parses text using `front-matter` package
- Sets resume object (to state) with `attributes` returned from
  front-matter, and adds a text field to the resume object.

Text component is added to the App, and wraps the text body in a
ReactMarkdown component.